fold createdocs into setup
